Updated data type of id from UUID to Int

+++ TextileTrader/Entity/Orders/OrderItem.swift
@@ -1,4 +1,4 @@
 struct OrderItem {
     let item: Item
-    let quantity: Int
+    let quantity: Double
 }
